==> 02_embeddings.ipynb <==
# Embedding Bug Reports with OpenAI <code>text-embedding-3-small</code>

<u><b>Full Procedure<u><b>

1. Load Dependencies and Dataset
- Necessary libraries are imported for data handling, pre-processing the text, and interaction with the OpenAI API. A custom text pre-processor and OpenAI client are initialized
- The dataset of resolved Mozilla Firefox bug reports is loaded into a pandas DataFrame

2. Preprocess the Text
- Concatenate the Summary and Description fields, as they will be used as input for the models in the next steps
- The bug report Concat field is pre-processed using the custom <code>PreProcessor</code> class to clean and prepare text for embedding

3. Tokenization and Embed the Text
- Each text entry is tokenized using the <code>tiktoken</code> library to ensure compatibility with OpenAI models and calculate token usage
- Each preprocessed comment is embedded using the OpenAI embedding API. Embeddings are stored along with their corresponding metadata

4. Save the Embeddings
- The computed embeddings and metadata are saved to a file for future use in tasks like classification, fine-tuning, clustering and semantic search

### Imports and Read Data

from typing import List
import pandas as pd
from text_preprocessing import PreProcessor

import tiktoken
from openai import OpenAI

client = OpenAI()
# Custom pre-processor class
pp = PreProcessor()

# Retrieve the bug reports from the request_data notebook
data = pd.read_csv("bug_reports_mozilla_firefox_resolved_fixed_comments.csv")

data.head()

data.Summary[0]

data.Description[0]

# Concatenate the Summary and Description fields in another column named Concat
data["Concat"] = ("Summary: " + data.Summary.str.strip() + "; Description: " + data.Description.str.strip())

data.head(2)

data.Concat[0]

### Pre-Process Text

# Pre-process the text using the custom class
data["Concat"] = data["Concat"].apply(pp.clean_text)

# Example of the cleaned text
data.Concat[0]

### Tokenization and Embed the Text

# Function to transform the text into embeddings using the model available in the OpenAI API
def get_embedding(text: str, model = "text-embedding-3-small", **kwargs) -> List[float]:
    # Replace newlines, which can negatively affect performance.
    text = text.replace("\n", " ")

    response = client.embeddings.create(input=[text], model = model, **kwargs)

    return response.data[0].embedding

embedding_model = "text-embedding-3-small"
embedding_encoding = "cl100k_base"
max_tokens = 8000  # The maximum for text-embedding-3-small is 8191

# Encode the text to check the lenght of each bug report
encoding = tiktoken.get_encoding(embedding_encoding)
data["N_tokens"] = data.Concat.apply(lambda x: len(encoding.encode(x)))

# There is one example with 16.877 words, so we will remove it from the dataset before creating the embeddings
max(data["N_tokens"])

data.drop(data[data["N_tokens"] == 16877].index, inplace = True)

data.reset_index(drop = True, inplace = True)

# Now, the example with the max lenght has 3.268 words, which is fine for the embedding model 
max(data["N_tokens"])

# Create a new column name Embeddings and apply the vectorization for each bug report text
# This may take a few minutes
data["Embeddings"] = data.Concat.apply(lambda x: get_embedding(x, model = embedding_model))

### Save Embeddings

# Save data to CSV
data.to_csv("bug_reports_mozilla_firefox_resolved_fixed_comments_embeddings.csv", index = None)
