==> 05_fine_tuning.ipynb <==
# 📘 Fine-Tuning with OpenAI GPT

<u><b>Full Procedure<u><b>
1. Prepare the Dataset
- Load and read a CSV file containing labeled bug reports, their description, and precomputed embeddings

2. Prepare Data to Fine-Tune <code>gpt-o-mini</code>
- Use <code>train_test_split</code> to divide the data into training and validation sets
- Transform the data into a JSONL format, required by OpenAI's fine-tuning API
- Upload and send the formatted JSONL file to OpenAI servers

3. Fine-Tune the Model 
- Start the fine-tuning training job using <code>gpt-o-mini</code>

4. Model Evaluation
- Use the fine-tuned against the <code>gpt-o-mini</code> without fine-tuning and classify new samples with both
- Compare predictions against true labels using standard classification metrics

### Imports and Read Data

import pandas as pd
import numpy as np
import json
from sklearn.model_selection import train_test_split
from sklearn.metrics import accuracy_score, precision_score, recall_score, f1_score, classification_report, confusion_matrix

from openai import OpenAI

client = OpenAI()

# Load the CSV file with the correct delimiter
data_path = "bug_reports_mozilla_firefox_resolved_fixed_comments_embeddings.csv"

data = pd.read_csv(data_path)

data.head(2)

# Drop all P3 examples
data = data[data["Priority"] != "P3"].reset_index(drop = True)

# Drop unnecessary columns
data_sub = data[["Concat", "Priority"]]

# Create a new subset of the data
data_sub.head()

# Mapping the 'label' column (Priority) to a more human-readable text
label_mapping = {"P1": "priority", "P2": "non-priority"}
data_sub["Priority"] = data_sub["Priority"].map(label_mapping)

# We can see that the class labels have been changed to 'priority' and 'non-priority'
data_sub.head()

### Prepare and Upload Data to OpenAI API

# Split the data into training and validation sets (80% train, 20% validation)
train_data, validation_data = train_test_split(data_sub, test_size = 0.2, random_state = 12345)

# Function to save data from CSV to JSON file with the message/role structure
def save_to_jsonl(data, output_file_path):
    jsonl_data = []
    for index, row in data.iterrows():
        jsonl_data.append({
            "messages": [
                {"role": "system", "content": "Given a bug report from Bugzilla, classify whether it is 'priority' or 'non-priority'."},
                {"role": "user", "content": row['Concat']},
                {"role": "assistant", "content": f"\"{row['Priority']}\""}
            ]
        })

    # Save to JSONL format
    with open(output_file_path, 'w') as f:
        for item in jsonl_data:
            f.write(json.dumps(item) + '\n')

# Save the training and validation sets to separate JSONL files
train_output_file_path = 'data_for_finetuning_prepared_train.jsonl' 
validation_output_file_path = 'data_for_finetuning_prepared_valid.jsonl'

save_to_jsonl(train_data, train_output_file_path)
save_to_jsonl(validation_data, validation_output_file_path)

# Check saved files
print(f"Training dataset save to: {train_output_file_path}")
print(f"Validation dataset save to: {validation_output_file_path}")

# Upload Dataset to OpenAI API
train_file = client.files.create(
  file = open(train_output_file_path, "rb"),
  purpose = "fine-tune"
)

valid_file = client.files.create(
  file = open(validation_output_file_path, "rb"),
  purpose = "fine-tune"
)

print(f"Training file Info: {train_file}")
print(f"Validation file Info: {valid_file}")

### Starting the Fine-Tuning Job

model = client.fine_tuning.jobs.create(
  training_file = train_file.id, 
  validation_file = valid_file.id,
  model = "gpt-4o-mini-2024-07-18", 
  hyperparameters = {
    "n_epochs": 5,
	"batch_size": 5
  }
)
job_id = model.id
status = model.status

print(f'Fine-tuning model with jobID: {job_id}.')
print(f"Training Response: {model}")
print(f"Training Status: {status}")

# Retrieve the state of a fine-tune
client.fine_tuning.jobs.retrieve(job_id)

# Assessing Fine-Tuned Model
result = client.fine_tuning.jobs.list()

# Retrieve the fine tuned model
fine_tuned_model = result.data[0].fine_tuned_model
print(fine_tuned_model)

# Check the response given from the fine-tuned model for a given bug report
completion = client.chat.completions.create(
  model = fine_tuned_model,
  messages=[
    {"role": "system", "content": "Given a bug report from Bugzilla, classify whether it is 'priority' or 'non-priority'."},
    {"role": "user", "content": "migrate preference experimental nimbus"}
  ]
)
print(completion.choices[0].message.content)

### Model Evaluation

# Predict function to pass the bug reports to the chat completion in order to get the classification response
def predict(test, model):
    
    y_pred = []
    categories = ["non-priority", "priority"]

    for index, row in test.iterrows():
        response = client.chat.completions.create(
            model = model,
            messages = [
                {
                    "role": "system",
                    "content": "Given a bug report from Bugzilla, classify whether it is 'priority' or 'non-priority'.",
                },
                {"role": "user", "content": row["Concat"]},
            ],
        )

        answer = response.choices[0].message.content

        # Determine the predicted category

        for category in categories:
            if category.lower() in answer.lower():
                y_pred.append(category)
                break
        else:
            y_pred.append("None")
            
    return y_pred

# Function to evaluate the model's performance
def evaluate(y_true, y_pred):
    labels = ["non-priority", "priority"]
    mapping = {label: idx for idx, label in enumerate(labels)}

    def map_func(x):
        return mapping.get(
            x, -1
        )  # Map to -1 if not found, but should not occur with correct data

    y_true_mapped = np.vectorize(map_func)(y_true)
    y_pred_mapped = np.vectorize(map_func)(y_pred)

    # Calculate accuracy

    accuracy = accuracy_score(y_true = y_true_mapped, y_pred = y_pred_mapped)
    print(f"Accuracy: {accuracy:.3f}")

    # Generate accuracy report for each class

    unique_labels = set(y_true_mapped)  # Get unique labels

    for label in unique_labels:
        label_indices = [
            i for i in range(len(y_true_mapped)) if y_true_mapped[i] == label
        ]
        label_y_true = [y_true_mapped[i] for i in label_indices]
        label_y_pred = [y_pred_mapped[i] for i in label_indices]
        label_accuracy = accuracy_score(label_y_true, label_y_pred)
        print(f"Accuracy for label {labels[label]}: {label_accuracy:.3f}")
        
    # Generate classification report

    class_report = classification_report(
        y_true = y_true_mapped,
        y_pred = y_pred_mapped,
        target_names = labels,
        labels = list(range(len(labels))),
    )
    print("\nClassification Report:")
    print(class_report)

    # Generate confusion matrix

    conf_matrix = confusion_matrix(
        y_true = y_true_mapped, y_pred = y_pred_mapped, labels = list(range(len(labels)))
    )
    print("\nConfusion Matrix:")
    print(conf_matrix)

# Evaluate model without fine-tuning
y_pred = predict(validation_data, "gpt-4o-mini-2024-07-18")
y_true = validation_data["Priority"]

evaluate(y_true, y_pred)

# Evaluate fine-tuned model
y_pred = predict(validation_data, fine_tuned_model)

evaluate(y_true, y_pred)
