==> 07_questions.ipynb <==
# Question answering using embeddings-based search

<u><b>Full procedure<b><u>

1. Prepare search data (once per document)

- Collect: We'll download a few hundred bug reports
- Chunk: Documents are split into short, mostly self-contained sections to be embedded
- Embed: Each section is embedded with the OpenAI API
- Store: Embeddings are saved (for large datasets, use a vector database)

2. Search (once per query)

- Given a user question, generate an embedding for the query from the OpenAI API
- Using the embeddings, rank the text sections by relevance to the query

3. Ask (once per query)

- Insert the question and the most relevant sections into a message to GPT
- Return GPT's answer

4. Answers
- Some example questions are made to the system
- The bug reports in which the answers are found can also be retrieved

### Imports and Read Data

import ast  
from scipy import spatial  
import pandas as pd 

from openai import OpenAI
import tiktoken 

# Create a list of models 
GPT_MODELS = ["gpt-4o", "gpt-4o-mini"]
# Embedding models
EMBEDDING_MODEL = "text-embedding-3-small"

client = OpenAI()

# Retrieve the bug reports from the request_data notebook
data_path = "bug_reports_mozilla_firefox_resolved_fixed_comments_embeddings.csv"

# The embeddings are already stored in the CSV, and need to be transformed from strings to lists
data = pd.read_csv(data_path)
data["Embeddings"] = data["Embeddings"].apply(ast.literal_eval)

### Search

Now we'll define a search function that:

- Takes a user query and a dataframe with text & embedding columns
- Embeds the user query with the OpenAI API
- Uses distance between query embedding and text embeddings to rank the texts
- Returns two lists:
1. The top N texts, ranked by relevance
2. Their corresponding relevance score

# Search function
def strings_ranked_by_relatedness(
    query: str,
    data: pd.DataFrame,
    relatedness_fn = lambda x, y: 1 - spatial.distance.cosine(x, y),
    top_n: int = 100
) -> tuple[list[str], list[float]]:
    
    """Returns a list of strings and relatednesses, sorted from most related to least."""
    query_embedding_response = client.embeddings.create(
        model = EMBEDDING_MODEL,
        input = query,
    )
    query_embedding = query_embedding_response.data[0].embedding
    strings_and_relatednesses = [
        (row["Concat"], relatedness_fn(query_embedding, row["Embeddings"]))
        for i, row in data.iterrows()
    ]
    strings_and_relatednesses.sort(key = lambda x: x[1], reverse = True)
    strings, relatednesses = zip(*strings_and_relatednesses)
    
    return strings[:top_n], relatednesses[:top_n]

# Examples
strings, relatednesses = strings_ranked_by_relatedness("User Interface (UI)", data, top_n = 5)
for string, relatedness in zip(strings, relatednesses):
    print(f"{relatedness = :.3f}")
    display(string)
    print("\n")

### Ask

With the search function above, we can now automatically retrieve relevant knowledge and insert it into messages to GPT. Below, we define a function ask that:

- Takes a user query
- Searches for text relevant to the query
- Stuffs that text into a message for GPT
- Sends the message to GPT
- Returns GPT's answer

def num_tokens(text: str, model: str = GPT_MODELS[0]) -> int:
    
    """Return the number of tokens in a string."""
    
    encoding = tiktoken.encoding_for_model(model)
    
    return len(encoding.encode(text))

def query_message(
    query: str,
    data: pd.DataFrame,
    model: str,
    token_budget: int
) -> str:
    
    """Return a message for GPT, with relevant source texts pulled from a dataframe."""
    
    strings, relatednesses = strings_ranked_by_relatedness(query, data)
    introduction = 'Use the bug reports from Bugzilla to answer the subsequent question. If the answer cannot be found, write "I could not find an answer."'
    question = f"\n\nQuestion: {query}"
    message = introduction
    
    for string in strings:
        next_bug = f'\n\nBug Report Summary:\n"""\n{string}\n"""'
        if (
            num_tokens(message + next_bug + question, model = model)
            > token_budget
        ):
            break
        else:
            message += next_bug
            
    return message + question

def ask(
    query: str,
    data: pd.DataFrame = data,
    model: str = GPT_MODELS[0],
    token_budget: int = 4096 - 500,
    print_message: bool = False,
) -> str:
    
    """Answers a query using GPT and a dataframe of relevant texts and embeddings."""
    
    message = query_message(query, data, model = model, token_budget = token_budget)
    
    if print_message:
        print(message)
        
    messages = [
        {"role": "system", "content": "You answer questions about the Bugzilla bug reports"},
        {"role": "user", "content": message},
    ]
    response = client.chat.completions.create(
        model = model,
        messages = messages,
        temperature = 0
    )
    response_message = response.choices[0].message.content
    
    return response_message

### Answers

print(ask('What is the most common issue in these bug reports?'))

print(ask('What are the issues related to User Interface (UI) in these bug reports?'))

print(ask('What are the issues related to User Interface (UI) in these bug reports?', print_message = True))

print(ask("About what the issues related to Firefox web browser talk about?"))

print(ask("What are the most common software components affected?"))

print(ask("What are the most common features affected?"))

print(ask("What words are used the most to describe the issues?"))
